Refactor tables again and seed DB

+++ database/migrations/0000_00_00_000080_create_processor_components_table.php
@@ -20,7 +20,7 @@ class CreateProcessorComponentsTable extends Migration
 
             // features
             $table->boolean('has_apu');
-            $table->float('speed');
+            $table->integer('speed');
 
             $table->integer('socket_id')->unsigned();
             $table->foreign('socket_id')->references('id')->on('sockets')->onDelete('restrict');
